Support getting more than 1k repos + Deduplication
GitHub limits the number of results that can be retrieved using their
search APIs to 1000. To work around that, I changed the code to make
multiple requests to the backend, where each subsequent request
limiting the results to star/fork count below the minimum value
obtained from the previous request.

I also added a dedulpication mechaism to the code. This is required
because, for some reason, GitHub API (or perhaps Octokit) is
returning deuplicate results.